requirements.txt
aiogram==3.4.1
fastapi==0.110.2
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.31
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
alembic==1.13.1
psycopg2-binary==2.9.9
greenlet>=0.4.18
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
python-multipart==0.0.9
Jinja2==3.1.4
geopy==2.4.1
transformers==4.40.0
torch==2.3.0
ruff==0.5.7
google-generativeai==0.8.5
